requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiomysql==0.2.0
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.8.0
asyncer==0.0.7
asyncpg==0.30.0
attrs==25.3.0
azure-core==1.34.0
azure-identity==1.23.0
azure-storage-blob==12.25.1
azure-storage-file-datalake==12.20.0
backoff==2.2.1
bcrypt==4.3.0
bidict==0.23.1
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
chainlit==2.5.5
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
colorama==0.4.6
cryptography==45.0.2
dataclasses-json==0.6.7
Deprecated==1.2.15
distro==1.9.0
et_xmlfile==2.0.0
faiss-cpu==1.11.0
fastapi==0.115.6
filelock==3.18.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.5.0
google-analytics-data==0.18.18
google-api-core==2.24.2
google-auth==2.40.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.4
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.8.2
kaleido==0.2.1
Lazify==0.4.0
literalai==0.1.201
lxml==5.4.0
mailjet-rest==1.3.4
MarkupSafe==3.0.2
marshmallow==3.24.1
mcp==1.9.1
monotonic==1.6
msal==1.32.3
msal-extensions==1.3.1
multidict==6.2.0
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.59.4
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.40.7
opentelemetry-instrumentation-anthropic==0.40.7
opentelemetry-instrumentation-bedrock==0.40.7
opentelemetry-instrumentation-chromadb==0.40.7
opentelemetry-instrumentation-cohere==0.40.7
opentelemetry-instrumentation-crewai==0.40.7
opentelemetry-instrumentation-google-generativeai==0.40.7
opentelemetry-instrumentation-groq==0.40.7
opentelemetry-instrumentation-haystack==0.40.7
opentelemetry-instrumentation-lancedb==0.40.7
opentelemetry-instrumentation-langchain==0.40.7
opentelemetry-instrumentation-llamaindex==0.40.7
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-marqo==0.40.7
opentelemetry-instrumentation-mcp==0.40.7
opentelemetry-instrumentation-milvus==0.40.7
opentelemetry-instrumentation-mistralai==0.40.7
opentelemetry-instrumentation-ollama==0.40.7
opentelemetry-instrumentation-openai==0.40.7
opentelemetry-instrumentation-pinecone==0.40.7
opentelemetry-instrumentation-qdrant==0.40.7
opentelemetry-instrumentation-replicate==0.40.7
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.40.7
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.40.7
opentelemetry-instrumentation-transformers==0.40.7
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.40.7
opentelemetry-instrumentation-watsonx==0.40.7
opentelemetry-instrumentation-weaviate==0.40.7
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.9
opentelemetry-util-http==0.52b1
packaging==23.2
pandas==2.2.3
pillow==11.2.1
plotly==5.24.1
polyline==2.0.2
posthog==3.25.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.27.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.18
python-pptx==1.0.2
python-socketio==5.12.1
python-telegram-bot==22.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.3.5
starlette==0.41.3
syncer==2.0.3
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
traceloop-sdk==0.40.7
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
uptrace==1.31.0
urllib3==2.3.0
uvicorn==0.25.0
watchfiles==0.20.0
websockets==15.0.1
wrapt==1.17.0
wsproto==1.2.0
XlsxWriter==3.2.3
yarl==1.19.0
zipp==3.21.0
